Elin_SukutsuArena/tools/pyproject.toml
[project]
name = "tools"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "openpyxl>=3.1.5",
    "pillow>=12.1.0",
]
